init: Next.js Pages Api Integration
