--- .prettierrc
@@ -1,6 +1,0 @@
-{
-  "singleQuote": true,
-  "semi": true,
-  "tabWidth": 2,
-  "trailingComma": "es5"
-}
