+++ src/component/resultsComponent/ResultsComponent.tsx
