--- src/main.tsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './components/app/App.tsx';
-import './index.css';
-import { Provider } from 'react-redux';
-import store from '../src/components/store/Store';
-
-ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
-  </React.StrictMode>,
-);
