+++ styles/searchHero.css
@@ -1,4 +1,4 @@
-/* SearchHero.css */
+/* searchHero.css */
 
 /* Light Theme */
 .search-hero.light-theme {
